[tune] Reduce sampling API clutter (#4739)
Adds some sugar for tune sampling API (for commonplace sampling idioms).
