+++ python/ray/tune/log_sync.py
@@ -19,7 +19,7 @@ import ray
 from ray.tune.cluster_info import get_ssh_key, get_ssh_user
 from ray.tune.error import TuneError
 from ray.tune.result import DEFAULT_RESULTS_DIR
-from ray.tune.suggest.variant_generator import function as tune_function
+from ray.tune.sample import function as tune_function
 
 logger = logging.getLogger(__name__)
 _log_sync_warned = False
