+++ python/ray/tune/suggest/__init__.py
@@ -1,16 +1,11 @@
 from ray.tune.suggest.search import SearchAlgorithm
 from ray.tune.suggest.basic_variant import BasicVariantGenerator
 from ray.tune.suggest.suggestion import SuggestionAlgorithm
-from ray.tune.suggest.variant_generator import grid_search, function, \
-    sample_from
+from ray.tune.suggest.variant_generator import grid_search
 
 __all__ = [
-    "SearchAlgorithm",
-    "BasicVariantGenerator",
-    "SuggestionAlgorithm",
-    "grid_search",
-    "function",
-    "sample_from",
+    "SearchAlgorithm", "BasicVariantGenerator", "SuggestionAlgorithm",
+    "grid_search"
 ]
